pyproject.toml
[tool.poetry]
name = "effective-mobile-trial-saucedemo"
version = "0.1.0"
description = ""
authors = ["Daria Egorova <spinoglaz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
